+++ src/redux/actions/cartActions.js
@@ -1,23 +1,23 @@
 export const addPizzaToCart = (pizzaObj) => ({
-   type: 'ADD_PIZZA_CART',
-   payload: pizzaObj,
+  type: 'ADD_PIZZA_CART',
+  payload: pizzaObj,
 });
 
 export const clearCart = () => ({
-   type: 'CLEAR_CART',
+  type: 'CLEAR_CART',
 });
 
 export const removeCartItem = (id) => ({
-   type: 'REMOVE_CART_ITEM',
-   payload: id,
+  type: 'REMOVE_CART_ITEM',
+  payload: id,
 });
 
 export const plusCartItem = (id) => ({
-   type: 'PLUS_CART_ITEM',
-   payload: id,
+  type: 'PLUS_CART_ITEM',
+  payload: id,
 });
 
 export const minusCartItem = (id) => ({
-   type: 'MINUS_CART_ITEM',
-   payload: id,
+  type: 'MINUS_CART_ITEM',
+  payload: id,
 });
