+++ src/redux/actions/index.js
@@ -1,0 +1,9 @@
+export { setSortBy, setCategory } from './filtersActions';
+export { setLoaded, fetchPizzas, setPizzas } from './pizzasActions';
+export {
+  addPizzaToCart,
+  clearCart,
+  removeCartItem,
+  plusCartItem,
+  minusCartItem,
+} from './cartActions';
