+++ src/redux/actions/pizzasActions.js
@@ -2,15 +2,22 @@ import axios from 'axios';
 
 export const setLoaded = (payload) => ({
   type: 'SET_LOADED',
-  payload
+  payload,
 });
+
 export const fetchPizzas = (sortBy, category) => (dispatch) => {
-  dispatch(setLoaded(false))
-  axios.get(`/pizzas?${category === null ?
-    '' : `category=${category}`}&_sort=${sortBy.type}&_order=${sortBy.order}`).then(({ data }) => {
+  dispatch(setLoaded(false));
+  axios
+    .get(
+      `/pizzas?${category === null ? '' : `category=${category}`}&_sort=${
+        sortBy.type
+      }&_order=${sortBy.order}`,
+    )
+    .then(({ data }) => {
       dispatch(setPizzas(data));
     });
 };
+
 export const setPizzas = (items) => ({
   type: 'SET_PIZZAS',
   payload: items,
